+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - Parse `ComplexType`s recursively
   - Rename `InstanceParser::parse_instance` to `InstanceParser::parse`
   - Rename `LinkbaseParser::parse_linkbase` to `InstanceParser::parse`
+  - Rename `SchemaParser::parse_schema` to `SchemaParser::parse`
 - removed
   - Remove `tuple_children` and `compositor` from `Concept`
   - Remove `context` for `XbrlError::FileOpen` and `XbrlError::FileRead`